commands/requirements.txt
pyyaml
requests
mangum
asgiref